refactor: StudyService 로직 분리 - 조회 / 생성, 수정, 삭제를 기준으로 서비스 클래스 분리
